Move input_schema.py to resource folder and use right import for fields
